models stay in external
